+++ packages/otel/src/bootstrap/main.ts
@@ -1,6 +1,11 @@
+import { type SpanProcessor } from '@opentelemetry/sdk-trace-node'
+import type { Instrumentation } from '@opentelemetry/instrumentation'
 import { GET_TRACER, SHUTDOWN_TRACERS } from '../constants.js'
 
-export const createTracerProvider = async (options: {
+export interface TracerProviderOptions {
+  /**
+   * The request headers (checked for presence of the `x-nf-enable-tracing` header)
+   */
   headers: Headers
   serviceName: string
   serviceVersion: string
@@ -8,7 +13,11 @@ export const createTracerProvider = async (options: {
   siteUrl: string
   siteId: string
   siteName: string
-}) => {
+  instrumentations?: (Instrumentation | Promise<Instrumentation>)[]
+  extraSpanProcessors?: (SpanProcessor | Promise<SpanProcessor>)[]
+}
+
+export const createTracerProvider = async (options: TracerProviderOptions) => {
   if (!options.headers.has('x-nf-enable-tracing')) {
     return
   }
@@ -17,9 +26,11 @@ export const createTracerProvider = async (options: {
   const runtimeVersion = nodeVersion.slice(1)
 
   const { Resource } = await import('@opentelemetry/resources')
-  const { NodeTracerProvider, BatchSpanProcessor } = await import('@opentelemetry/sdk-trace-node')
+  const { NodeTracerProvider, SimpleSpanProcessor } = await import('@opentelemetry/sdk-trace-node')
 
-  const { NetlifySpanExporter } = await import('./netlify_span_exporter.js')
+  const { registerInstrumentations } = await import('@opentelemetry/instrumentation')
+
+  const { NetlifySpanExporter } = await import('../exporters/netlify.js')
 
   const resource = new Resource({
     'service.name': options.serviceName,
@@ -34,10 +45,20 @@ export const createTracerProvider = async (options: {
 
   const nodeTracerProvider = new NodeTracerProvider({
     resource,
-    spanProcessors: [new BatchSpanProcessor(new NetlifySpanExporter())],
+    spanProcessors: [
+      new SimpleSpanProcessor(new NetlifySpanExporter()),
+      ...(await Promise.all(options.extraSpanProcessors ?? [])),
+    ],
   })
 
   nodeTracerProvider.register()
+
+  const instrumentations = await Promise.all(options.instrumentations ?? [])
+
+  registerInstrumentations({
+    instrumentations,
+    tracerProvider: nodeTracerProvider,
+  })
 
   const { trace } = await import('@opentelemetry/api')
   const { SugaredTracer } = await import('@opentelemetry/api/experimental')