+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [2.5.3](https://github.com/netlify/primitives/compare/vite-plugin-v2.5.2...vite-plugin-v2.5.3) (2025-07-29)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.5.2 to 4.5.3
+    * @netlify/dev-utils bumped from ^4.1.0 to ^4.1.1
+
+## [2.5.2](https://github.com/netlify/primitives/compare/vite-plugin-v2.5.1...vite-plugin-v2.5.2) (2025-07-28)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.5.1 to 4.5.2
+
+## [2.5.1](https://github.com/netlify/primitives/compare/vite-plugin-v2.5.0...vite-plugin-v2.5.1) (2025-07-24)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.5.0 to 4.5.1
+
 ## [2.5.0](https://github.com/netlify/primitives/compare/vite-plugin-v2.4.6...vite-plugin-v2.5.0) (2025-07-23)